Ecommerce/requirement.txt
asgiref==3.2.8
certifi==2020.6.20
chardet==3.0.4
Django==3.0.7
django-cors-headers==3.4.0
django-oauth2-provider==0.2.6.1
djangorestframework==3.11.0
httpie==2.2.0
idna==2.9
Pygments==2.6.1
pytz==2020.1
requests==2.24.0
shortuuid==1.0.1
sqlparse==0.3.1
Unipath==1.1
urllib3==1.25.9